Added a check for the user if they are deleting posts not belonging to them

+++ routers/posts.py
@@ -113,6 +113,12 @@ async def delete_post(
             detail=f"The post with id of {_id} was not found"
         )
 
+    if _post.user_id != _user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to perform this operation"
+        )
+
     _db.delete(_post)
     _db.commit()
